refactor(app-wide): delete unused files and rename and restructure to improve consistency
Vestigial and duplicate recipe detail page and home page section components were removed. Files were
renamed and moved to conform to the @/assets, @/features, @/pages structure in use in our project.
Within the assets folder, files were organized.
