+++ src/App.jsx
@@ -1,15 +1,7 @@
 import Spa from "./routes";
 import makeServer from "./mirageServer/server";
 import { Routes, Route } from "react-router-dom";
-import {
-  Home,
-  Posts,
-  PostDetail,
-  About,
-  Search,
-  RecipeDetailsPage,
-  Components,
-} from "@/pages";
+import { Home, About, Search, RecipeDetails, Components } from "@/pages";
 // Please keep on when possible to avoid using up our Tasty API free quota
 // All api calls will be intercepted and fulfilled by the mirage server.
 const USE_MIRAGE_API = true;
@@ -24,11 +16,9 @@ function App() {
       <Route element={<Spa />}>
         <Route path="/" element={<Home />} />
         <Route path="/about" element={<About />} />
-        <Route path="/posts" element={<Posts />} />
-        <Route path="/posts/:postId" element={<PostDetail />} />
         <Route path="/search" element={<Search />} />
         <Route path="/recipes">
-          <Route path=":recipeId" element={<RecipeDetailsPage />} />
+          <Route path=":recipeId" element={<RecipeDetails />} />
         </Route>
         <Route path="/components" element={<Components />} />
       </Route>
