+++ src/assets/brand/food/index.js
@@ -1,0 +1,5 @@
+import chickenWings from "./ChickenWings.png";
+import chocoPBChips from "./ChocolateChipPB.png";
+import spicyChickenPizza from "./SpicyChickenPizza.png";
+
+export { chickenWings, chocoPBChips, spicyChickenPizza };
