+++ src/assets/brand/index.js
@@ -1,0 +1,5 @@
+import BellPeppers from "./bell-peppers/BellPeppers";
+import YumiWithSpatula from "./yumi-with-spatula/YumiWithSpatula";
+import YumiWithSpoon from "./yumi-with-spoon/YumiWithSpoon";
+
+export { BellPeppers, YumiWithSpatula, YumiWithSpoon };
