+++ src/assets/brand/logo/index.js
@@ -1,0 +1,4 @@
+import colorLogo from "./yumyumyes-logo-m.png";
+import creamLogo from "./yumyumyes-logo-lightcream.svg";
+
+export { colorLogo, creamLogo };
