+++ src/assets/brand/logo/yumyumyes-logo-m.png
