+++ src/assets/brand/swooshes/TopWaveSm.png
