+++ src/assets/index.js
@@ -1,21 +1,8 @@
-import colorLogo from "./brand/yumyumyes-logo-m.png";
-import creamLogo from "./brand/yumyumyes-logo-lightcream.svg";
-import BellPeppers from "./brand/bell-peppers/BellPeppers";
-import SpicyChickenPizza from "./HomePage/SpicyChickenPizza.png";
-import ChickenWings from "./HomePage/ChickenWingsfood.png";
-import ChocoPBChip from "./HomePage/ChocolateChipPB.png";
 import FeatureRecipe from "./RecipePhoto.png";
-import github from "./github.png";
 
-export {
-  github,
-  colorLogo,
-  creamLogo,
-  SpicyChickenPizza,
-  ChickenWings,
-  ChocoPBChip,
-  BellPeppers,
-  FeatureRecipe,
-};
+export { FeatureRecipe };
 
+export * from "./brand";
+export * from "./brand/food";
+export * from "./brand/logo";
 export * from "./icons";
