+++ src/features/recipes/index.js
@@ -1,3 +1,4 @@
 export * from "./api";
 export * from "./components/recipe-list";
 export * from "./components/feature-of-the-day";
+export * from "./components/recipe-link-card";
