+++ src/features/ui/index.js
@@ -8,4 +8,3 @@ export * from "./Button";
 export * from "./Picture";
 export * from "./LoadingState";
 export * from "../HomePage";
-export * from "./RecipeLinkCard";
