--- src/pages/components/Dev.jsx
@@ -1,13 +1,0 @@
-import Components from "./Components";
-import Icons from "./Icons";
-
-const DevPage = () => {
-  return (
-    <div>
-      <Components />
-      <Icons />
-    </div>
-  );
-};
-
-export default DevPage;
