--- src/pages/home/svgWave.jsx
@@ -1,14 +1,0 @@
-const SvgComponent = (props) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    fill="none"
-    viewBox="0 0 1280 110"
-    {...props}
-  >
-    <path
-      fill="#FFC945"
-      d="m0 15.788 42.667 12.798c42.666 12.606 128 38.49 213.333 51.096C341.333 92.48 426.667 92.48 512 82.27c85.333-10.401 170.667-30.533 256-35.805 85.333-5.273 170.667 5.272 256 0 85.33-5.273 170.67-25.405 213.33-35.806L1280 .45v109.435H0V15.788Z"
-    />
-  </svg>
-);
-export default SvgComponent;
