--- src/pages/recipes/RecipeDetailPage.jsx
@@ -1,9 +1,0 @@
-const RecipesDetailPage = () => {
-  return (
-    <div className="p-10">
-      <h1 className="text-center">Recipes Detail Page</h1>
-    </div>
-  );
-};
-
-export default RecipesDetailPage;
